Update deprecation comment on SystemdCgroups
This field was deprecated in favor of the same one in Runtime.Options,
but we're no longer looking at that field either, so document it as
doubly deprecated now.

Deprecating the one in Runtime.Options is a bit more tricky, since it's
in a protobuf that's part of an official API. Need to figure out what to
do about that one.

Signed-off-by: Filipe Brandenburger <filbranden@google.com>