introduce multi-tenancy via managed namespaces
CRI clients are containerd clients. containerd supports multi-tenancy
via containerd namespaces. Clients interact with resources in non-default
namespaces by specifying a namespace name in the client invocation that
is curried to the server via gRPC. This means that the basic building
blocks for a client-directed (aka honor-system) multi-tenancy are ready
to be leveraged.

Introducing "managed namespaces" for containerd/cri. This change-set is
largely concerned with allowing for multi-tenant CRI to coexist and
enabling interactions with CRI resources in non-default (aka not
"k8s.io") namespaces by NOT overriding the client-provided namespace.
This means that the default CRI namespace, "k8s.io", is only enforced
for clients that have not specified a namespace. A CRI namespace is one
that is labeled with "io.cri-containerd=managed". Invoking the CRI with
a namespace that either does not exist or isn't properly labeled will
return an "unmanaged namespace" error.

How was this implemented?

- the plugin initialization aspect of `criService` and the check/logging of `instrumentedService` have been merged into `criServiceManager`
- the `criServiceManager` maintains an event exchange that is subscribed to containerd events (filtering on task and image events, as before)
- the `criServiceManager` manages `criService` instances in a sync.Map keyed by namespace name
- all `criService` instances are subscribed to the `criServiceManager` event exchange filtering on their namespace name
- all `criService` instances are currently created and started on NamespaceCreate/NamespaceUpdate having the label "io.cri-containerd == managed"
  - managed `criService` instances are stopped on NamespaceDelete or NamespaceUpdate when "io.cri-containerd != managed"
- the `DeferContext` and `NamespacedContext` in ./pkg/containerd/util have been changed to accept a namespace name
- the default CRI namespace, k8s.io, is configured, as before, by the CRI plugin TOML
- all other namespaces are configured by TOML that is the standard, aka "version 1", emitted by serializing a `config.PluginConfig`
  - located at `${containerd.root_dir}/io.containerd.grpc.v1.cri/namespaces/${namespace.name}/config.toml`
  - this config is serialized for all namespaces, including k8s.io, at the same (parameterized) location
  - non-default namespaces lacking a config file will be started with a "default" having:
    - `cni.conf_dir = /opt/cri/${namespace.name}/etc/cni/net.d` instead of the usual default `/etc/cni/net.d`

Signed-off-by: Jacob Blain Christen <jacob@rancher.com>